Added cross-validation to Squires' fit
The script loads simulations used to fit the Squires' data. For each model (with its best parameter), we try to fit (by leaving one observation out, which is repeated for every one of the 48 observations) its predictions with other models (by looping over their free parameter and choosing the best one). Reporting R2 of this fit allow to assess how identifiable are the models.